remove statsd code to separate file
